Made pdf converter general, pointing it towards the correct folders

+++ detecht_backend/detecht_api/detecht_converter/pdf_converter.py
@@ -2,36 +2,38 @@ import PyPDF2
 import json
 import os
 import re
-import unicodedata
 
-#Titta på textract som alternativ
 def convert_pdf_to_json():
 
-    # get the PDF path and read the file
-    file = 'detecht_api/static/pdf/BOAML_report_2016.pdf'
+    # get all PDFs  in the path
+    path = "detecht_api/static/pdf"
+    with os.scandir(path) as it:
+        for entry in it:
+            if entry.name.endswith(".pdf") and entry.is_file():
+                print("Currently scanning: " + entry.name)
+                file = entry.path
 
-    #file = open(file, "r")
-    read_pdf = PyPDF2.PdfFileReader(file, strict=False)
-    all_pages = {}
-    #file.close()
+                # read the current pdf
+                read_pdf = PyPDF2.PdfFileReader(file, strict=False)
+                all_pages = {}
 
-    # iterate the page numbers
-    for page in range(read_pdf.getNumPages()):
-        data = read_pdf.getPage(page)
-        #page_mode = read_pdf.getPageMode()
-        page_text = data.extractText()
+                # iterate all pages
+                all_text = ""
+                for page in range(read_pdf.getNumPages()):
+                    data = read_pdf.getPage(page)
+                    page_text = data.extractText()
+                    modified_text = re.sub("\n", "", page_text)
+                    all_text = all_text + modified_text
 
-        modified_text = re.sub(" \n", "", page_text)
-        #modified_text = re.sub("\n","",page_text)
+                all_pages["All_text"] = all_text
 
-        #modified_text = re.compile(modified_text, re.UNICODE)
-        #all_pages[page] = modified_text
-        all_pages[page] = modified_text
-
-    # create a JSON string from the dictionary
-    json_data = json.dumps(all_pages)
-    with open('BOAML_report_2016.json', 'w+') as outfile:
-        json.dump(all_pages, outfile)
+                # creates a JSON file
+                #json_data = json.dumps(all_pages)
+                filename = entry.name
+                filename = filename[:-4]
+                filepath = 'detecht_api/static/json/' + filename + '.json'
+                with open(filepath,'w+') as outfile:
+                    json.dump(all_pages, outfile)
